+++ partiql-beamline-serde/src/lib.rs
@@ -1,2 +1,3 @@
+pub mod ddl;
 pub mod kollider;
 pub mod serde;